+++ src/base/base.js
@@ -57,8 +57,6 @@
         // Base生命周期
         initialize: function(config) {
             config = config || {};
-            this._parseEventsFromConfig(config);
-            this._parseEventsFromInstance();
             this._initAttrs(config);
         },
         destroy: function() {
@@ -71,16 +69,6 @@
             this.destroy = function() {};
         },
 
-        // 解析prototype上得 _onChangeX, _beforeMethod, _afterMethod 
-        _parseEventsFromInstance: function() {
-            parseEvents(this, $.extend({}, this), INSTANCE_EVENT_PATTERN);
-        },
-
-        // 解析配置上的onChangeX, beforeMethod, afterMethod 
-        _parseEventsFromConfig: function(config) {
-            parseEvents(this, config, CONFIG_EVENT_PATTERN);
-        },
-
         /***************************** Events ******************************/
 
         on: function(events, callback, context) {
@@ -88,7 +76,7 @@
 
             if(!callback) return this;
             
-            cache = (this._events = this._events || {});
+            cache = (this.__events = this.__events || {});
             events = events.split(EVENT_SPLITTER);
             while(event = events.shift()) {
                  cache[event] = cache[event] || [];
@@ -101,11 +89,11 @@
         // this.off('switch') 清除全部switch事件的处理函数
         // this.off('switch', 'fun1'); 清除switch事件的fun1处理函数
         off: function(events, callback) {
-            var cache = this._events, event;
+            var cache = this.__events, event;
 
             // 全部为空，则清除全部handler
             if(!(events || callback)) {
-                delete this._events;
+                delete this.__events;
                 return this;
             }
             events = events.split(EVENT_SPLITTER);
@@ -131,7 +119,7 @@
         // this.trigger('switch change', [args1, args2]);
         // @return true/false
         trigger: function(events) {
-            var cache = this._events, event, 
+            var cache = this.__events, event, 
                 me = this, 
                 returnValue = true;
 
@@ -140,7 +128,7 @@
             events = events.split(EVENT_SPLITTER);
             while(event = events.shift()) {
                 var handlers = cache[event];
-                var ev = new CustEvent(me, event, {a: 1, b: 2});
+                var ev = new CustEvent(me, event);
                 if(handlers) {
                     for(var i = 0, len = handlers.length; i < len; i += 2) {
                         var ctx = handlers[i + 1] || me;
@@ -230,12 +218,15 @@
 
                 // 设置子属性，如set('a.b.c') 
                 if(isSubAttr) {
-                    var newValue = $.extend({}, prevVal.value);
-                    var subAttr = getSubproperty(newValue, path.slice(1, -1));  
-                    if(subAttr == ATTRIBUTE.INVALID_VALUE || !$.isPlainObject(subAttr)) {
+                    // 获得a.b，如果a.b为undefined，或者不是plain object。则set失败
+                    var subAttr = getProperty(prevVal.value, path.slice(1, -1).join('.'));
+                    if(subAttr == undefined || !$.isPlainObject(subAttr)) {
                         setStatus = false;
                         return;
                     }
+
+                    var newValue = $.extend({}, prevVal.value);
+                    subAttr = getProperty(newValue, path.slice(1, -1).join('.'));
                     subAttr[path[path.length - 1]] = val;
                     val = newValue;
                 }
@@ -259,6 +250,7 @@
                 now[attrName].value = val;
                 if(!options.silent) {
                     me.trigger('change:' + attrName, [me.get(attrName), prevVal, name, options.data]); 
+                    me.trigger('change:*', [me.get(attrName), prevVal, name, options.data]); 
                 }
             });
 
@@ -301,7 +293,7 @@
                 }
 
                 // 根据path返回
-                val =  getSubproperty(val, path.slice(1)) ;
+                val = getProperty(val, path.slice(1).join('.'));
                 return val == ATTRIBUTE.INVALID_VALUE ? undefined : val;
             }
         },
@@ -343,36 +335,13 @@
             // 调用setter初始化value
             $.each(config, function(name, val) {
                 if(curAttrs[name].setter) {
-                    me.set(name, val, null, true);
+                    me.set(name, val, {silent: true});
                 } 
             });
         }
 
 
     });
-
-    /**************************** Events parser helpers *******************************/
-
-    // 解析prototype和config中的onChangeAttr, beforeMethod, afterMethod等事件
-    function parseEvents(widget, obj, pattern) {
-        $.each(obj, function(name, fun) {
-            var m = name.match(pattern);
-            if(!m || !fun ||  $.type(fun) != 'function') return;
-            if(m[1] == 'onChange') {
-                var attrName = firstLetterLc(m[2]);
-                widget.on('change:' + attrName, fun); 
-            }
-            else {
-                var when = m[1];        // before/after
-                var methodName = firstLetterLc(m[2]);
-
-                if($.type(widget[methodName] == 'function')) {
-                    widget[when](methodName, fun);
-                }
-            }
-            delete obj[name];
-        }); 
-    }
 
 
     /**************************** Attribute Helpers *******************************/
@@ -416,6 +385,22 @@
             return ATTRIBUTE.INVALID_VALUE;
         }
     }
+
+    // eg. getProperty(obj, 'a.b.c')会返回obj.a.b.c
+    function getProperty(obj, prop) {
+        var keys = prop ? prop.split(".") : [];
+        var ret = obj;
+
+        for(var i = 0, len = keys.length; i < len; i++) {
+            if(ret == undefined) {
+                return;
+            }
+            ret = ret[keys[i]];
+        }
+
+        return ret;
+    }
+
 
     /***************************** Apspect helpers ******************************/